refactor : 패키지 구조 변경 & 플러그인 오류 제거
IDE에서 layGround.java 정보를 인식하지 못하는 문제 발생
- IDE 초기화하였지만 동작 실패
- 패키지 분리 & 네이밍 변경 후 정상 동작 제출을 위해 긴급 변경
- 테스트 & 동작 확인 완료

editorConfig, checkStyle 관련 오류 발생(빌드 & 테스트는 됨)
- 양털깎기 대신 동작 될 수 있게 플러그인 해제
