+++ build.gradle
@@ -1,8 +1,6 @@
 plugins {
     id 'java'
     id 'eclipse'
-    id 'org.ec4j.editorconfig' version '0.0.3'
-//    id 'checkstyle'
 }
 version = '1.0.0'
 sourceCompatibility = 1.8
@@ -10,20 +8,6 @@ sourceCompatibility = 1.8
 repositories {
     mavenCentral()
 }
-
-// .editconfig
-editorconfig {
-    excludes = ['build']
-}
-check.dependsOn editorconfigCheck
-
-//checkstyle {
-//    maxWarnings = 0 // 규칙이 어긋나는 코드가 하나라도 있을 경우 빌드 fail을 내고 싶다면 이 선언을 추가한다.
-//    configFile = file("${rootDir}/config/checkstyle/naver-chestyle-rules.xml")
-//    configProperties = ["suppressionFile" : "${rootDir}/config/checkstyle/naver-checkstyle-suppressions.xml"]
-//    toolVersion ="8.24"  // checkstyle 버전 8.24 이상 선언
-//}
-
 
 // utf-8 인코딩
 compileJava.options.encoding = 'UTF-8'
