+++ src/main/java/baseball/Program.java
@@ -5,11 +5,11 @@ import baseball.ui.UiSystem;
 import baseball.ui.impl.ConsoleUiSystem;
 
 public class Program {
-	public static void main(String[] args) {
+    public static void main(String[] args) {
 
-		UiSystem uiSystem = new ConsoleUiSystem();
-		NumberBaseBall numberBaseBall = new NumberBaseBall(uiSystem);
-		numberBaseBall.run();
-	}
+        UiSystem uiSystem = new ConsoleUiSystem();
+        NumberBaseBall numberBaseBall = new NumberBaseBall(uiSystem);
+        numberBaseBall.run();
+    }
 
 }
