+++ src/main/java/baseball/game/NumberBaseBall.java
@@ -1,51 +1,52 @@
 package baseball.game;
 
-import baseball.game.playground.PlayGround;
-import baseball.game.playground.enums.SystemCommand;
+import baseball.game.enums.Command;
+import baseball.game.ground.Ground;
+import baseball.game.referee.Referee;
 import baseball.ui.UiSystem;
 
 public class NumberBaseBall {
 
-	private static final String SYSTEM_COMMENT = String.format("게임을 새로 시작하려면 %s, 종료하려면 %s를 입력하세요",
-		SystemCommand.AGAIN.getCommand(),
-		SystemCommand.EXIT.getCommand());
+    private static final String GUIDE_MESSAGE = String.format("게임을 새로 시작하려면 %s, 종료하려면 %s를 입력하세요",
+            Command.AGAIN.getCommand(),
+            Command.EXIT.getCommand());
 
-	private final UiSystem uiSystem;
-	private final PlayGround playGround;
+    private final UiSystem uiSystem;
+    private final Ground ground;
+    private final Referee referee;
 
-	public NumberBaseBall(UiSystem uiSystem) {
-		this.uiSystem = uiSystem;
-		this.playGround = new PlayGround();
-	}
+    public NumberBaseBall(UiSystem uiSystem) {
+        this.uiSystem = uiSystem;
+        this.ground = new Ground(uiSystem);
+        this.referee = new Referee();
+    }
 
-	public void run() {
+    public void run() {
 
-		boolean isRunning = true;
-		while (isRunning) {
+        boolean isRunning = true;
+        while (isRunning) {
 
-			playGround.playBall(uiSystem);
-			playGround.update();
+            referee.playBall(ground);
+            isRunning = onEventInput();
+        }
+        uiSystem.display("게임을 종료합니다");
+    }
 
-			isRunning = onSystemUpdate();
-		}
-		uiSystem.display("게임을 종료합니다");
-	}
+    private boolean onEventInput() {
 
-	private boolean onSystemUpdate() {
+        while (true) {
+            uiSystem.display(GUIDE_MESSAGE);
 
-		while (true) {
-			uiSystem.display(SYSTEM_COMMENT);
+            String inputString = uiSystem.gameInput().trim().toLowerCase();
+            Command cmd = Command.parse(inputString);
+            if (Command.canAgain(cmd)) {
+                return true;
+            }
 
-			String inputString = uiSystem.gameInput().trim().toLowerCase();
-			SystemCommand cmd = SystemCommand.parse(inputString);
-			if (SystemCommand.canAgain(cmd)) {
-				return true;
-			}
-
-			if (SystemCommand.canExit(cmd)) {
-				return false;
-			}
-			uiSystem.display(SystemCommand.UNKNOWN.name());
-		}
-	}
+            if (Command.canExit(cmd)) {
+                return false;
+            }
+            uiSystem.display(Command.UNKNOWN.name());
+        }
+    }
 }
