+++ src/main/java/baseball/game/gameball/validatitor/condition/ConditionPolicy.java
@@ -1,0 +1,6 @@
+package baseball.game.gameball.validatitor.condition;
+
+public interface ConditionPolicy {
+
+    boolean isSatisfied(String inputString);
+}
