--- src/main/java/baseball/game/playground/gameball/validatitor/condition/ConditionPolicy.java
@@ -1,6 +1,0 @@
-package baseball.game.playground.gameball.validatitor.condition;
-
-public interface ConditionPolicy {
-
-	boolean isSatisfied(String inputString);
-}
