+++ src/main/java/baseball/ui/impl/ConsoleUiSystem.java
@@ -1,24 +1,24 @@
 package baseball.ui.impl;
 
-import java.util.Scanner;
-
 import baseball.ui.UiSystem;
+
+import java.util.Scanner;
 
 /**
  * 콘솔 ui를 사용하기위한 객체
  */
 public class ConsoleUiSystem implements UiSystem {
-	private final Scanner scanner = new Scanner(System.in);
+    private final Scanner scanner = new Scanner(System.in);
 
-	// check validation
-	@Override
-	public String gameInput() {
-		return scanner.nextLine();
-	}
+    // check validation
+    @Override
+    public String gameInput() {
+        return scanner.nextLine();
+    }
 
-	@Override
-	public void display(String data) {
-		System.out.println(data);
-	}
+    @Override
+    public void display(String data) {
+        System.out.println(data);
+    }
 
 }
