+++ src/main/java/baseball/ui/impl/TestUiSystem.java
@@ -7,34 +7,34 @@ import baseball.ui.UiSystem;
  */
 public class TestUiSystem implements UiSystem {
 
-	// test only
-	private String gameInput;
-	private String displayData;
+    // test only
+    private String gameInput;
+    private String displayData;
 
-	public String getGameInput() {
-		return gameInput;
-	}
+    public String getGameInput() {
+        return gameInput;
+    }
 
-	public void setGameInput(String gameInput) {
-		this.gameInput = gameInput;
-	}
+    public void setGameInput(String gameInput) {
+        this.gameInput = gameInput;
+    }
 
-	public String getDisplayData() {
-		return displayData;
-	}
+    public String getDisplayData() {
+        return displayData;
+    }
 
-	public void setDisplayData(String displayData) {
-		this.displayData = displayData;
-	}
+    public void setDisplayData(String displayData) {
+        this.displayData = displayData;
+    }
 
-	@Override
-	public String gameInput() {
-		return gameInput;
-	}
+    @Override
+    public String gameInput() {
+        return gameInput;
+    }
 
-	@Override
-	public void display(String data) {
-		displayData = data;
-	}
+    @Override
+    public void display(String data) {
+        displayData = data;
+    }
 
 }
